renamed map-to-string as MapToString
